fix(croquis): split semantic snapshot modules
